--- a/enterprise/server/remote_execution/containers/podman/BUILD
+++ b/enterprise/server/remote_execution/containers/podman/BUILD
@@ -27,12 +27,12 @@ go_library(
         "//server/util/log",
         "//server/util/lru",
         "//server/util/networking",
+        "//server/util/proto",
         "//server/util/random",
         "//server/util/retry",
         "//server/util/status",
         "//server/util/tracing",
         "@com_github_masterminds_semver_v3//:semver",
         "@com_github_prometheus_client_golang//prometheus",
-        "@org_golang_google_protobuf//proto",
     ],
 )