--- a/enterprise/server/remote_execution/containers/sandbox/BUILD
+++ b/enterprise/server/remote_execution/containers/sandbox/BUILD
@@ -18,8 +18,8 @@ go_library(
         "//proto:runner_go_proto",
         "//server/interfaces",
         "//server/util/log",
+        "//server/util/proto",
         "//server/util/status",
-        "@org_golang_google_protobuf//proto",
     ],
 )