<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**:
Benchmark test reveals the marshaling ExecutionSummary with vtproto is 9% slower, though our current code is not marshaling ExecutionSummary. 

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                                     │      Old      │                 New                  │
                                     │    sec/op     │    sec/op     vs base                │
Marshal/pbName=ExecutionSummary-24      18.22µ ± 33%   19.88µ ±  6%   +9.09% (p=0.001 n=10)
Marshal/pbName=RunnerKey-24             4.974µ ±  7%   3.004µ ±  8%  -39.61% (p=0.000 n=10)
Marshal/pbName=RunnerPoolState-24       245.2µ ±  3%   106.7µ ±  3%  -56.51% (p=0.000 n=10)
Marshal/pbName=Platform-24              4.087µ ± 11%   2.571µ ±  5%  -37.07% (p=0.000 n=10)
Marshal/pbName=ExecutionTask-24         24.47µ ±  7%   12.97µ ±  4%  -46.98% (p=0.000 n=10)
Marshal/pbName=WorkResponse-24          339.1n ±  6%   211.6n ± 11%  -37.59% (p=0.000 n=10)
Marshal/pbName=FileNode-24              5.181µ ± 29%   3.514µ ±  4%  -32.18% (p=0.000 n=10)
Unmarshal/pbName=Platform-24           12.181µ ± 22%   8.667µ ± 15%  -28.85% (p=0.000 n=10)
Unmarshal/pbName=ExecutionTask-24       85.06µ ±  5%   60.63µ ±  3%  -28.73% (p=0.000 n=10)
Unmarshal/pbName=WorkResponse-24        479.2n ±  6%   355.6n ± 15%  -25.79% (p=0.002 n=10)
Unmarshal/pbName=FileNode-24           13.211µ ± 14%   9.488µ ± 11%  -28.18% (p=0.000 n=10)
Unmarshal/pbName=ExecutionSummary-24    54.28µ ±  4%   45.03µ ±  6%  -17.04% (p=0.000 n=10)
Unmarshal/pbName=RunnerKey-24           15.74µ ± 27%   10.10µ ± 14%  -35.86% (p=0.000 n=10)
Unmarshal/pbName=RunnerPoolState-24     761.8µ ±  5%   513.9µ ± 10%  -32.54% (p=0.000 n=10)
geomean                                 14.41µ         9.702µ        -32.68%

                                     │      Old      │                   New                   │
                                     │      B/s      │      B/s        vs base                 │
Marshal/pbName=ExecutionSummary-24     526.3Mi ± 49%    482.4Mi ±  6%    -8.34% (p=0.001 n=10)
Marshal/pbName=RunnerKey-24            649.6Mi ±  7%   1075.7Mi ±  7%   +65.60% (p=0.000 n=10)
Marshal/pbName=RunnerPoolState-24      114.0Mi ±  3%    262.0Mi ±  3%  +129.92% (p=0.000 n=10)
Marshal/pbName=Platform-24             1.238Gi ± 10%    1.968Gi ±  5%   +58.96% (p=0.000 n=10)
Marshal/pbName=ExecutionTask-24        727.4Mi ±  7%   1372.0Mi ±  4%   +88.62% (p=0.000 n=10)
Marshal/pbName=WorkResponse-24         87.20Mi ±  6%   140.09Mi ± 10%   +60.65% (p=0.000 n=10)
Marshal/pbName=FileNode-24             570.6Mi ± 22%    840.6Mi ±  4%   +47.30% (p=0.000 n=10)
Unmarshal/pbName=Platform-24           385.8Mi ± 18%    542.5Mi ± 17%   +40.61% (p=0.000 n=10)
Unmarshal/pbName=ExecutionTask-24      234.4Mi ±  6%    328.9Mi ±  4%   +40.31% (p=0.000 n=10)
Unmarshal/pbName=WorkResponse-24       61.71Mi ±  5%    83.13Mi ± 13%   +34.72% (p=0.002 n=10)
Unmarshal/pbName=FileNode-24           65.77Mi ± 12%    91.57Mi ± 10%   +39.24% (p=0.000 n=10)
Unmarshal/pbName=ExecutionSummary-24   28.48Mi ±  4%    34.33Mi ±  6%   +20.54% (p=0.000 n=10)
Unmarshal/pbName=RunnerKey-24          300.2Mi ± 36%    468.1Mi ± 14%   +55.92% (p=0.000 n=10)
Unmarshal/pbName=RunnerPoolState-24    309.8Mi ±  5%    459.2Mi ± 11%   +48.25% (p=0.000 n=10)
geomean                                236.8Mi          351.8Mi         +48.57%

                                     │     Old      │                  New                   │
                                     │     B/op     │     B/op      vs base                  │
Marshal/pbName=ExecutionSummary-24     12.87Ki ± 0%   17.12Ki ± 0%  +33.10% (p=0.000 n=10)
Marshal/pbName=RunnerKey-24            2.802Ki ± 0%   2.802Ki ± 0%        ~ (p=1.000 n=10) ¹
Marshal/pbName=RunnerPoolState-24      162.1Ki ± 0%   162.1Ki ± 0%        ~ (p=0.303 n=10)
Marshal/pbName=Platform-24             2.493Ki ± 0%   2.493Ki ± 0%        ~ (p=1.000 n=10) ¹
Marshal/pbName=ExecutionTask-24        20.13Ki ± 0%   20.14Ki ± 0%   +0.08% (p=0.000 n=10)
Marshal/pbName=WorkResponse-24           32.00 ± 0%     32.00 ± 0%        ~ (p=1.000 n=10) ¹
Marshal/pbName=FileNode-24             2.742Ki ± 0%   2.758Ki ± 0%   +0.57% (p=0.000 n=10)
Unmarshal/pbName=Platform-24           7.763Ki ± 0%   7.763Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/pbName=ExecutionTask-24      58.95Ki ± 0%   58.95Ki ± 0%        ~ (p=0.526 n=10)
Unmarshal/pbName=WorkResponse-24         112.0 ± 0%     112.0 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/pbName=FileNode-24           8.007Ki ± 0%   8.007Ki ± 0%        ~ (p=1.000 n=10)
Unmarshal/pbName=ExecutionSummary-24   32.50Ki ± 0%   32.50Ki ± 0%        ~ (p=1.000 n=10)
Unmarshal/pbName=RunnerKey-24          9.172Ki ± 0%   9.172Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/pbName=RunnerPoolState-24    558.5Ki ± 0%   558.5Ki ± 0%   +0.00% (p=0.008 n=10)
geomean                                7.468Ki        7.625Ki        +2.11%
¹ all samples are equal

                                     │      Old      │                   New                   │
                                     │   allocs/op   │  allocs/op   vs base                    │
Marshal/pbName=ExecutionSummary-24      1.000 ± 0%     60.000 ± 0%  +5900.00% (p=0.000 n=10)
Marshal/pbName=RunnerKey-24             1.000 ± 0%      1.000 ± 0%          ~ (p=1.000 n=10) ¹
Marshal/pbName=RunnerPoolState-24       1.000 ± 0%      1.000 ± 0%          ~ (p=1.000 n=10) ¹
Marshal/pbName=Platform-24              0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
Marshal/pbName=ExecutionTask-24         1.000 ± 0%      3.000 ± 0%   +200.00% (p=0.000 n=10)
Marshal/pbName=WorkResponse-24          1.000 ± 0%      1.000 ± 0%          ~ (p=1.000 n=10) ¹
Marshal/pbName=FileNode-24              1.000 ± 0%      3.000 ± 0%   +200.00% (p=0.000 n=10)
Unmarshal/pbName=Platform-24            149.0 ± 0%      149.0 ± 0%          ~ (p=1.000 n=10) ¹
Unmarshal/pbName=ExecutionTask-24      1.014k ± 0%     1.014k ± 0%          ~ (p=1.000 n=10) ¹
Unmarshal/pbName=WorkResponse-24        2.000 ± 0%      2.000 ± 0%          ~ (p=1.000 n=10) ¹
Unmarshal/pbName=FileNode-24            154.0 ± 0%      154.0 ± 0%          ~ (p=1.000 n=10) ¹
Unmarshal/pbName=ExecutionSummary-24    576.0 ± 0%      576.0 ± 0%          ~ (p=1.000 n=10) ¹
Unmarshal/pbName=RunnerKey-24           174.0 ± 0%      174.0 ± 0%          ~ (p=1.000 n=10) ¹
Unmarshal/pbName=RunnerPoolState-24    10.75k ± 0%     10.75k ± 0%          ~ (p=1.000 n=10) ¹
geomean                                            ²                  +56.74%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```